+++ lora/CHANGELOG.md
@@ -33,6 +33,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - rename `Interrupt` to `IRQ` for type and driver methods
 - dio methods bound by `DIO` trait
 - irq methods bound by `IRQ` trait
+- `set_power_amplifier` to `set_tx_power`
+- `PARamp` to `TxPowerRamp`
+- `pa_ramp()` to `tx_power_ramp()`
+- `set_pa_ramp()` to `set_tx_power_ramp()`
 
 ### Fixed
 
@@ -46,6 +50,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `assert!` from driver
 - unused `RxStatus` enum
 - `DioXSignal` enums
+- `PowerAmplifier` struct
 
 ## [0.1.1] - 2026-04-23